fix(M): rename to Auth.ts
