+++ frontend/src/interfaces/models/Auth.ts
@@ -1,0 +1,16 @@
+export interface CreateUserRequest {
+  full_name: string;
+  email: string;
+  phone: string;
+  password: string;
+  // code: string;
+}
+
+export interface SendVerificationCodeRequest {
+  phone_number: string;
+}
+
+export interface VerifyCodeRequest {
+  phone_number: string;
+  code: string;
+}
