--- frontend/src/interfaces/models/CreateUserRequest.ts
@@ -1,7 +1,0 @@
-export interface CreateUserRequest {
-  full_name: string;
-  email: string;
-  phone: string;
-  password: string;
-  // code: string;
-}
